Make the markup service stateless
