+++ src/elife_profile/modules/custom/elife_article/elife_article.module
@@ -236,7 +236,7 @@ function elife_article_markup_service() {
     $markup_service = $factory();
   }
 
-  return clone $markup_service;
+  return $markup_service;
 }
 
 /**
