+++ src/elife_profile/modules/custom/elife_article/infrastructure/src/eLife/MarkupService/MarkupNotFound.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace eLife\MarkupService;
+
+use RuntimeException;
+
+class MarkupNotFound extends RuntimeException {
+}
