+++ src/elife_profile/modules/custom/elife_article/infrastructure/src/eLife/MarkupService/MarkupService.php
@@ -4,28 +4,11 @@ namespace eLife\MarkupService;
 
 interface MarkupService {
   /**
-   * @param string $article_id
-   * @param string $section
-   */
-  public function addSectionQuery($article_id, $section);
-
-  /**
-   * @param string $article_id
-   * @param string $doi
-   */
-  public function addDoiQuery($article_id, $doi);
-
-
-  public function submitQuery();
-
-  /**
-   * @param string $response
-   */
-  public function processResponse($response);
-
-  /**
+   * @param Query $query
+   *
    * @return string
+   *
+   * @throws MarkupNotFound
    */
-  public function output();
+  public function query(Query $query);
 }
-
